backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pyodbc
pydantic
python-dotenv
python-multipart
